backend/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
sqlalchemy==1.4.23
python-dotenv==0.19.0
twilio==7.16.0
python-multipart==0.0.5